+++ src/api/resources/users/resources/rewards/client/requests/index.ts
@@ -2,3 +2,4 @@ export type { GetBatchesByPlacementRequest } from "./GetBatchesByPlacementReques
 export type { GetLocationsByUserRequest } from "./GetLocationsByUserRequest.js";
 export type { GetOffersByPlacementRequest } from "./GetOffersByPlacementRequest.js";
 export type { GetOffersByUserRequest } from "./GetOffersByUserRequest.js";
+export type { GetPlacementContentRequest } from "./GetPlacementContentRequest.js";